api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pgvector
transformers
torch
torchvision
torchaudio
openai